prueba

+++ src/main/java/com/example/demo/dao/IUsuarioDAO.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.dto.Usuario;
 
-public interface IUsuarioDAO extends JpaRepository<Usuario, Long> {
+public interface IUsuarioDAO extends JpaRepository<Usuario, Integer> {
 
 	Usuario findByUsername(String username);
 }
